Move login panel related files to own namespace and tidy up class nesting
