+++ osu.Game/Overlays/LoginOverlay.cs
@@ -5,17 +5,17 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Graphics;
-using osu.Game.Overlays.Settings.Sections.General;
 using osuTK.Graphics;
 using osu.Framework.Graphics.Shapes;
 using osu.Game.Graphics.Containers;
 using osu.Game.Graphics.Cursor;
+using osu.Game.Overlays.Login;
 
 namespace osu.Game.Overlays
 {
     public class LoginOverlay : OsuFocusedOverlayContainer
     {
-        private LoginSettings settingsSection;
+        private LoginPanel panel;
 
         private const float transition_time = 400;
 
@@ -50,7 +50,7 @@ namespace osu.Game.Overlays
                             AutoSizeEasing = Easing.OutQuint,
                             Children = new Drawable[]
                             {
-                                settingsSection = new LoginSettings
+                                panel = new LoginPanel
                                 {
                                     Padding = new MarginPadding(10),
                                     RequestHide = Hide,
@@ -75,17 +75,17 @@ namespace osu.Game.Overlays
         {
             base.PopIn();
 
-            settingsSection.Bounding = true;
+            panel.Bounding = true;
             this.FadeIn(transition_time, Easing.OutQuint);
 
-            GetContainingInputManager().ChangeFocus(settingsSection);
+            GetContainingInputManager().ChangeFocus(panel);
         }
 
         protected override void PopOut()
         {
             base.PopOut();
 
-            settingsSection.Bounding = false;
+            panel.Bounding = false;
             this.FadeOut(transition_time);
         }
     }
